+++ cli.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2020 Lars Eric Scheidler
+Copyright 2021 Lars Eric Scheidler
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -19,33 +19,15 @@ package main
 import (
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"os"
 	"strings"
 
+	"github.com/lscheidler/letsencrypt-deploy/config"
 	"github.com/lscheidler/letsencrypt-deploy/hook"
 	"github.com/lscheidler/letsencrypt-deploy/hook/aws/sns"
 	"github.com/lscheidler/letsencrypt-deploy/hook/exec"
 )
-
-type Domains []string
-
-// String is the method to format the flag's value, part of the flag.Value interface.
-// The String method's output will be used in diagnostics.
-func (i *Domains) String() string {
-	return fmt.Sprint(*i)
-}
-
-// Set is the method to set the flag value, part of the flag.Value interface
-// Set's argument is a string to be parsed to set the flag.
-// It's a comma-separated list, so we split it.
-func (i *Domains) Set(value string) error {
-	for _, t := range strings.Split(value, ",") {
-		*i = append(*i, t)
-	}
-	return nil
-}
 
 type Hooks []*hook.Hook
 
@@ -82,19 +64,13 @@ func (i *Hooks) Set(value string) error {
 }
 
 var (
-	delay                 int
-	domains               Domains
-	dynamodbTableName     *string
-	dynamodbTableNameFlag string
-	email                 string
-	hooks                 Hooks
-	outputLocation        string
-	passphraseFile        string
-	passphrase            string
-	prefix                string
+	configFile string
+	hooks      Hooks
 )
 
 const (
+	configFileUsage             = "set config file"
+	configFileDefaultVal        = "/etc/letsencrypt-deploy/config.json"
 	delayUsage                  = "deploy certificates with a delay after creation (days)"
 	delayDefaultVal             = 0
 	domainUsage                 = "domains"
@@ -102,16 +78,20 @@ const (
 	dynamodbTableNameUsage      = "dynamodb table name"
 	emailDefaultVal             = ""
 	emailUsage                  = "account email"
+	fortiosUsage                = "deploy certificate on fortios firewall"
+	fortiosDefaultVal           = false
+	fortiosBaseUrlUsage         = "fortios base url"
+	fortiosBaseUrlDefaultVal    = ""
+	localUsage                  = "deploy certificate on local machine"
+	localDefaultVal             = false
 	outputLocationDefaultVal    = "/etc/ssl/private"
 	outputLocationUsage         = "output location for certificates"
-	passphraseFileDefaultVal    = ""
-	passphraseFileUsage         = "passphrase file"
 	prefixDefaultVal            = "letsencrypt."
 	prefixUsage                 = "file prefix for letsencrypt certificates"
 	hooksUsage                  = "run hook after certificates has updated"
 )
 
-func parseArgs() {
+func parseArgs() *config.Config {
 	flag.Usage = func() {
 		fmt.Fprintf(flag.CommandLine.Output(), "Usage of %s (%s):\n", os.Args[0], version)
 		flag.PrintDefaults()
@@ -131,43 +111,56 @@ hooks:
 		fmt.Fprintf(flag.CommandLine.Output(), "%s\n", examples)
 	}
 
-	flag.IntVar(&delay, "delay", delayDefaultVal, delayUsage)
+	var (
+		domains config.Domains
+	)
+
+	conf := &config.Config{}
+	flag.StringVar(&configFile, "configFile", configFileDefaultVal, configFileUsage)
+	flag.StringVar(&configFile, "c", configFileDefaultVal, configFileUsage)
+	if delay := flag.Int("delay", delayDefaultVal, delayUsage); *delay != delayDefaultVal {
+		conf.Delay = delay
+	}
 	flag.Var(&domains, "domain", domainUsage)
 	flag.Var(&domains, "d", domainUsage)
-	flag.StringVar(&dynamodbTableNameFlag, "dynamodbTableName", dynamodbTableNameDefaultVal, dynamodbTableNameUsage)
-	flag.StringVar(&dynamodbTableNameFlag, "t", dynamodbTableNameDefaultVal, dynamodbTableNameUsage)
-	flag.StringVar(&email, "email", emailDefaultVal, emailUsage)
-	flag.StringVar(&email, "e", emailDefaultVal, emailUsage)
-	flag.StringVar(&outputLocation, "outputLocation", outputLocationDefaultVal, outputLocationUsage)
-	flag.StringVar(&outputLocation, "o", outputLocationDefaultVal, outputLocationUsage)
-	flag.StringVar(&passphraseFile, "passphraseFile", passphraseFileDefaultVal, passphraseFileUsage)
-	flag.StringVar(&passphraseFile, "p", passphraseFileDefaultVal, passphraseFileUsage)
-	flag.StringVar(&prefix, "prefix", prefixDefaultVal, prefixUsage)
+	if len(domains) > 0 {
+		conf.Domains = domains
+	}
+	if dynamodbTableName := flag.String("dynamodbTableName", dynamodbTableNameDefaultVal, dynamodbTableNameUsage); *dynamodbTableName != dynamodbTableNameDefaultVal {
+		conf.DynamodbTableName = dynamodbTableName
+	}
+	if dynamodbTableName := flag.String("t", dynamodbTableNameDefaultVal, dynamodbTableNameUsage); *dynamodbTableName != dynamodbTableNameDefaultVal {
+		conf.DynamodbTableName = dynamodbTableName
+	}
+	if fortiosDeployment := flag.Bool("fortios", fortiosDefaultVal, fortiosUsage); *fortiosDeployment != fortiosDefaultVal {
+		conf.Fortios = *fortiosDeployment
+	}
+	if fortiosBaseUrl := flag.String("fortios-base-url", fortiosBaseUrlDefaultVal, fortiosBaseUrlUsage); *fortiosBaseUrl != fortiosBaseUrlDefaultVal {
+		conf.FortiosBaseUrl = fortiosBaseUrl
+	}
+	if email := flag.String("email", emailDefaultVal, emailUsage); *email != emailDefaultVal {
+		conf.Email = email
+	}
+	if email := flag.String("e", emailDefaultVal, emailUsage); *email != emailDefaultVal {
+		conf.Email = email
+	}
+	if localDeployment := flag.Bool("local", localDefaultVal, localUsage); *localDeployment != localDefaultVal {
+		conf.Local = *localDeployment
+	}
+	if outputLocation := flag.String("outputLocation", outputLocationDefaultVal, outputLocationUsage); *outputLocation != outputLocationDefaultVal {
+		conf.OutputLocation = outputLocation
+	}
+	if outputLocation := flag.String("o", outputLocationDefaultVal, outputLocationUsage); *outputLocation != outputLocationDefaultVal {
+		conf.OutputLocation = outputLocation
+	}
+	if prefix := flag.String("prefix", prefixDefaultVal, prefixUsage); *prefix != prefixDefaultVal {
+		conf.Prefix = prefix
+	}
 
 	flag.Var(&hooks, "hook", hooksUsage)
 	flag.Var(&hooks, "H", hooksUsage)
 
 	flag.Parse()
 
-	if len(dynamodbTableNameFlag) > 0 {
-		dynamodbTableName = &dynamodbTableNameFlag
-	}
-
-	if len(email) == 0 {
-		log.Fatal("Option -email must be set.")
-	}
-
-	if len(domains) == 0 {
-		log.Fatal("Option -domain must be set.")
-	}
-
-	if len(passphraseFile) == 0 {
-		log.Fatal("Option -passphraseFile must be set.")
-	}
-
-	if dat, err := ioutil.ReadFile(passphraseFile); err != nil {
-		log.Fatal(err)
-	} else {
-		passphrase = strings.TrimSpace(string(dat))
-	}
+	return conf
 }