+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.3 (2021-01-07)
+
+- introduce config file
+- added provider interface
+- added fortios firewall provider
+
 # v0.2 (2020-11-25)
 
 - added -delay to deploy renewed certificates <days> after renewing